Working arrow that points to north, refactor sensor code for list view and arrow into its own centralized OrientationManager.  TODO - point arrow to nearest stop location, update distance text.
