+++ onebusaway-android/src/main/res/layout/glass_arrival_card_list.xml
@@ -38,13 +38,27 @@
                 android:id="@+id/stop_name_footer"
                 android:textSize="18dp"
                 android:textStyle="bold"
-                android:layout_width="fill_parent"
+                android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_toRightOf="@id/bus_icon_footer"/>
+                android:ellipsize="end"
+                android:singleLine="true"
+                android:layout_toRightOf="@id/bus_icon_footer"
+                android:layout_toLeftOf="@+id/dist_to_stop"/>
+
+        <TextView
+                android:id="@+id/dist_to_stop"
+                android:textSize="14dp"
+                android:text="54 ft"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginLeft="10px"
+                android:layout_toLeftOf="@+id/arrow"/>
+
         <com.joulespersecond.oba.glass.ArrowView
                 android:id="@+id/arrow"
-                android:layout_width="50px"
-                android:layout_height="50px"
+                android:layout_width="30px"
+                android:layout_height="30px"
+                android:layout_marginLeft="10px"
                 android:layout_alignParentRight="true"/>
     </RelativeLayout>
